requirements.txt
pika==1.3.0
python-dotenv==0.20.0
requests==2.28.1
